chore: generalize binary downloading
