+++ crates/acvm_backend_barretenberg/src/lib.rs
@@ -5,8 +5,11 @@ use std::path::PathBuf;
 
 mod bb;
 mod cli;
+mod download;
 mod proof_system;
 mod smart_contract;
+
+use download::download_backend;
 
 const BACKENDS_DIR: &str = ".nargo/backends";
 
@@ -33,7 +36,7 @@ fn assert_binary_exists(backend: &Backend) -> PathBuf {
     let binary_path = backend.binary_path();
 
     if !binary_path.is_file() {
-        bb::download_bb_binary(&binary_path)
+        download_backend(&bb::get_bb_download_url(), &binary_path)
     }
     binary_path
 }
